fix(suite-native): navigation
