+++ suite-native/module-accounts/src/index.ts
@@ -1,3 +1,4 @@
 export { AccountsStackNavigator } from './navigation/AccountsStackNavigator';
 export { AccountSettingsScreen } from './screens/AccountSettingsScreen';
 export { AccountDetailScreen } from './screens/AccountDetailScreen';
+export { AccountsScreen } from './screens/AccountsScreen';
