+++ suite-native/module-home/package.json
@@ -18,6 +18,7 @@
         "@suite-native/atoms": "workspace:*",
         "@suite-native/formatters": "workspace:*",
         "@suite-native/graph": "workspace:*",
+        "@suite-native/module-accounts": "workspace:*",
         "@suite-native/module-settings": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@trezor/icons": "workspace:*",
